refactor(protocol): extract intent snapshot provenance
